guess-game-be/pyproject.toml
[project]
name = "guess-game-be"
version = "0.1.0"
description = ""
authors = [
    {name = "Gokul Sudheesh",email = "gokulsudheesh1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-core (>=0.3.63,<0.4.0)",
    "langchain-nvidia-ai-endpoints (>=0.3.10,<0.4.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
]

[tool.poetry]
packages = [{include = "guess_game_be", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
